// app/shared/models/participante-campanha-franquia.ts
export class ParticipanteCampanhaFranquia{
    idParticipantesCampanhaFranquia: number;
    idPessoaJuridica: number;
    dataTermoAceite: Date;
    idCampanhaFranquia:number;
    titulo: string;
    dataInicio: Date;
    dataFim: Date;
    dataLimiteAceite: Date;
    status: string;
    ano: number;
    semestre: number;
    dataLimiteComprovacao: Date;
    dataLimiteNotaDebito: Date;
    dataLimitePagamento: Date;
    dataLimiteAprovacao: Date;
    vcpHolding: number;
    vcpFranquia: number;
    vcpTotal: number;
    aceiteTermo: number;
    termoCondicoes: string;
    urlArquivoPolitica: string;
    codigoFranquia: number;
}